pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "youtwo"
version = "0.1.0"
authors = [{name = "Tan"}]
description = "Memory layer and RAG service"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "mcp",
    "smolagents",
    "requests",
    "python-dotenv",
]

[project.optional-dependencies]
frontend = [
    "gradio",
]
deployment = [
    "modal",
]
graph = [
    "langgraph",
    "langchain_core", 
    "matplotlib",
    "networkx",
]

all = [
    "youtwo[frontend]",
    "youtwo[deployment]",
    "youtwo[graph]"
]

# Ruff config

[tool.ruff]
exclude = [
    "**/__pycache__",
    "**/.git", 
    "**/.venv",
    "**/node_modules",
    "**/dist",
    "**/backend",
    "**/project-website"
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C", "N", "Q", "S", "T", "B", "PL"]
ignore = ["E501"]
isort = { known-first-party = ["youtwo"] }

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80
